Player now can acess Cabin
In order for the player to see stats, cabin has been implemented along with minor new additions.
Game keeps track on player duel count, won duel count, won money ammount etc. Information of all
of these can be found in the cabin. Cabin also has been added as a graphic into Town.cpp :)

Alongside player can now see oponents stas during the selection menu.

+++ application/code/game/Modifiers/Modifiers.cpp
@@ -5,13 +5,13 @@
 namespace Modifiers
 {
 
-    double duelReward(const double& enemyWeaponDmg, const double& enemyWeaponAccuracy, 
+    int duelReward(const double& enemyWeaponDmg, const double& enemyWeaponAccuracy, 
                       const double& distance, const int& playerCharisma,
                       const double& baseReward)
     {
-        double enemyMod = enemyWeaponDmg + enemyWeaponAccuracy * 100;
-        double distanceMod = distance * 2;
-        double charismaMod = 1 + (static_cast<double>(playerCharisma) / 10);
+        int enemyMod = enemyWeaponDmg + (enemyWeaponAccuracy * 100);
+        int distanceMod = distance * 2;
+        int charismaMod = 1 + (static_cast<double>(playerCharisma) / 10);
 
         return (baseReward + enemyMod + distanceMod) * charismaMod;
     }
